+++ targets/appengine/tests/test/testTests.py
@@ -1,8 +1,10 @@
-import unittest
-
 from google.appengine.api import memcache
 from google.appengine.ext import db
 from google.appengine.ext import testbed
+
+import testCase
+
+import mop.store.PhoneNumber
 
 class TestModel(db.Model):
   """A model class used for testing."""
@@ -23,18 +25,10 @@ def GetEntityViaMemcache(entity_key):
     memcache.set(entity_key, entity)
   return entity
 
-class DemoTestCase(unittest.TestCase):
-  def setUp(self):
-    # First, create an instance of the Testbed class.
-    self.testbed = testbed.Testbed()
-    # Then activate the testbed, which prepares the service stubs for use.
-    self.testbed.activate()
-    # Next, declare which service stubs you want to use.
-    self.testbed.init_datastore_v3_stub()
-    self.testbed.init_memcache_stub()
-
-  def tearDown(self):
-    self.testbed.deactivate()
+class DemoTestCase(testCase.TestCase):
+  def _initStubs(self, testbed):
+    testbed.init_datastore_v3_stub()
+    testbed.init_memcache_stub()
 
   def testInsertEntity(self):
     TestModel().put()
